@@ -1,4 +1,5 @@
 # This example requires the 'members' privileged intent to use the Member converter.
+# This example also requires the 'message_content' privileged intent to function.
 
 import typing
 
@@ -7,6 +8,7 @@ from discord.ext import commands
 
 intents = discord.Intents.default()
 intents.members = True
+intents.message_content = True
 
 bot = commands.Bot('!', intents=intents)
 
@@ -28,8 +30,9 @@ async def userinfo(ctx: commands.Context, user: discord.User):
     # and can do the following:
     user_id = user.id
     username = user.name
-    avatar = user.avatar_url
-    await ctx.send('User found: {} -- {}\n{}'.format(user_id, username, avatar))
+    avatar = user.display_avatar.url
+    await ctx.send(f'User found: {user_id} -- {username}\n{avatar}')
+
 
 @userinfo.error
 async def userinfo_error(ctx: commands.Context, error: commands.CommandError):
@@ -37,6 +40,7 @@ async def userinfo_error(ctx: commands.Context, error: commands.CommandError):
     # so we handle this in this error handler:
     if isinstance(error, commands.BadArgument):
         return await ctx.send('Couldn\'t find that user.')
+
 
 # Custom Converter here
 class ChannelOrMemberConverter(commands.Converter):
@@ -70,18 +74,18 @@ class ChannelOrMemberConverter(commands.Converter):
         # If the value could not be converted we can raise an error
         # so our error handlers can deal with it in one place.
         # The error has to be CommandError derived, so BadArgument works fine here.
-        raise commands.BadArgument('No Member or TextChannel could be converted from "{}"'.format(argument))
-
+        raise commands.BadArgument(f'No Member or TextChannel could be converted from "{argument}"')
 
 
 @bot.command()
 async def notify(ctx: commands.Context, target: ChannelOrMemberConverter):
     # This command signature utilises the custom converter written above
     # What will happen during command invocation is that the `target` above will be passed to
-    # the `argument` parameter of the `ChannelOrMemberConverter.convert` method and 
+    # the `argument` parameter of the `ChannelOrMemberConverter.convert` method and
     # the conversion will go through the process defined there.
 
-    await target.send('Hello, {}!'.format(target.name))
+    await target.send(f'Hello, {target.name}!')
+
 
 @bot.command()
 async def ignore(ctx: commands.Context, target: typing.Union[discord.Member, discord.TextChannel]):
@@ -95,9 +99,10 @@ async def ignore(ctx: commands.Context, target: typing.Union[discord.Member, dis
 
     # To check the resulting type, `isinstance` is used
     if isinstance(target, discord.Member):
-        await ctx.send('Member found: {}, adding them to the ignore list.'.format(target.mention))
-    elif isinstance(target, discord.TextChannel): # this could be an `else` but for completeness' sake.
-        await ctx.send('Channel found: {}, adding it to the ignore list.'.format(target.mention))
+        await ctx.send(f'Member found: {target.mention}, adding them to the ignore list.')
+    elif isinstance(target, discord.TextChannel):  # this could be an `else` but for completeness' sake.
+        await ctx.send(f'Channel found: {target.mention}, adding it to the ignore list.')
+
 
 # Built-in type converters.
 @bot.command()
@@ -109,5 +114,6 @@ async def multiply(ctx: commands.Context, number: int, maybe: bool):
     if maybe is True:
         return await ctx.send(number * 2)
     await ctx.send(number * 5)
+
 
 bot.run('token')
