@@ -1,9 +1,12 @@
+# This example requires the 'message_content' privileged intent to function.
+
 import discord
+
 
 class MyClient(discord.Client):
     async def on_ready(self):
-        print('Connected!')
-        print('Username: {0.name}\nID: {0.id}'.format(self.user))
+        print(f'Logged in as {self.user} (ID: {self.user.id})')
+        print('------')
 
     async def on_message(self, message):
         if message.content.startswith('!deleteme'):
@@ -14,8 +17,12 @@ class MyClient(discord.Client):
             await message.channel.send('Goodbye in 3 seconds...', delete_after=3.0)
 
     async def on_message_delete(self, message):
-        fmt = '{0.author} has deleted the message: {0.content}'
-        await message.channel.send(fmt.format(message))
+        msg = f'{message.author} has deleted the message: {message.content}'
+        await message.channel.send(msg)
 
-client = MyClient()
+
+intents = discord.Intents.default()
+intents.message_content = True
+
+client = MyClient(intents=intents)
 client.run('token')
