@@ -1,12 +1,13 @@
+# This example requires the 'message_content' privileged intent to function.
+
 import discord
 import random
 import asyncio
 
+
 class MyClient(discord.Client):
     async def on_ready(self):
-        print('Logged in as')
-        print(self.user.name)
-        print(self.user.id)
+        print(f'Logged in as {self.user} (ID: {self.user.id})')
         print('------')
 
     async def on_message(self, message):
@@ -25,12 +26,16 @@ class MyClient(discord.Client):
             try:
                 guess = await self.wait_for('message', check=is_correct, timeout=5.0)
             except asyncio.TimeoutError:
-                return await message.channel.send('Sorry, you took too long it was {}.'.format(answer))
+                return await message.channel.send(f'Sorry, you took too long it was {answer}.')
 
             if int(guess.content) == answer:
                 await message.channel.send('You are right!')
             else:
-                await message.channel.send('Oops. It is actually {}.'.format(answer))
+                await message.channel.send(f'Oops. It is actually {answer}.')
 
-client = MyClient()
+
+intents = discord.Intents.default()
+intents.message_content = True
+
+client = MyClient(intents=intents)
 client.run('token')
