@@ -1,3 +1,6 @@
+# This example requires the 'message_content' privileged intent to function.
+
+
 import random
 
 import discord
@@ -16,8 +19,8 @@ class MyContext(commands.Context):
             await self.message.add_reaction(emoji)
         except discord.HTTPException:
             # sometimes errors occur during this, for example
-            # maybe you dont have permission to do that
-            # we dont mind, so we can just ignore them
+            # maybe you don't have permission to do that
+            # we don't mind, so we can just ignore them
             pass
 
 
@@ -27,25 +30,30 @@ class MyBot(commands.Bot):
         # subclass to the super() method, which tells the bot to
         # use the new MyContext class
         return await super().get_context(message, cls=cls)
-        
 
-bot = MyBot(command_prefix='!')
+
+intents = discord.Intents.default()
+intents.message_content = True
+
+bot = MyBot(command_prefix='!', intents=intents)
+
 
 @bot.command()
 async def guess(ctx, number: int):
-    """ Guess a random number from 1 to 6. """
+    """Guess a random number from 1 to 6."""
     # explained in a previous example, this gives you
     # a random number from 1-6
     value = random.randint(1, 6)
     # with your new helper function, you can add a
     # green check mark if the guess was correct,
-    # or a red cross mark if it wasnt
+    # or a red cross mark if it wasn't
     await ctx.tick(number == value)
 
-# important: you shouldnt hard code your token
-# these are very important, and leaking them can 
+
+# IMPORTANT: You shouldn't hard code your token
+# these are very important, and leaking them can
 # let people do very malicious things with your
-# bot. try to use a file or something to keep
-# them private, and dont commit it to GitHub
+# bot. Try to use a file or something to keep
+# them private, and don't commit it to GitHub
 token = "your token here"
 bot.run(token)
