@@ -3,17 +3,20 @@ import typing
 import discord
 from discord.ext import commands
 
-bot = commands.Bot(command_prefix=commands.when_mentioned, description="Nothing to see here!")
+intents = discord.Intents.default()
 
-# the `hidden` keyword argument hides it from the help command. 
+bot = commands.Bot(command_prefix=commands.when_mentioned, description="Nothing to see here!", intents=intents)
+
+# the `hidden` keyword argument hides it from the help command.
 @bot.group(hidden=True)
 async def secret(ctx: commands.Context):
     """What is this "secret" you speak of?"""
     if ctx.invoked_subcommand is None:
         await ctx.send('Shh!', delete_after=5)
 
+
 def create_overwrites(ctx, *objects):
-    """This is just a helper function that creates the overwrites for the 
+    """This is just a helper function that creates the overwrites for the
     voice/text channels.
 
     A `discord.PermissionOverwrite` allows you to determine the permissions
@@ -26,10 +29,7 @@ def create_overwrites(ctx, *objects):
 
     # a dict comprehension is being utilised here to set the same permission overwrites
     # for each `discord.Role` or `discord.Member`.
-    overwrites = {
-        obj: discord.PermissionOverwrite(view_channel=True)
-        for obj in objects
-    }
+    overwrites = {obj: discord.PermissionOverwrite(view_channel=True) for obj in objects}
 
     # prevents the default role (@everyone) from viewing the channel
     # if it isn't already allowed to view the channel.
@@ -40,15 +40,16 @@ def create_overwrites(ctx, *objects):
 
     return overwrites
 
+
 # since these commands rely on guild related features,
 # it is best to lock it to be guild-only.
 @secret.command()
 @commands.guild_only()
 async def text(ctx: commands.Context, name: str, *objects: typing.Union[discord.Role, discord.Member]):
-    """This makes a text channel with a specified name 
+    """This makes a text channel with a specified name
     that is only visible to roles or members that are specified.
     """
-    
+
     overwrites = create_overwrites(ctx, *objects)
 
     await ctx.guild.create_text_channel(
@@ -57,6 +58,7 @@ async def text(ctx: commands.Context, name: str, *objects: typing.Union[discord.
         topic='Top secret text channel. Any leakage of this channel may result in serious trouble.',
         reason='Very secret business.',
     )
+
 
 @secret.command()
 @commands.guild_only()
@@ -70,8 +72,9 @@ async def voice(ctx: commands.Context, name: str, *objects: typing.Union[discord
     await ctx.guild.create_voice_channel(
         name,
         overwrites=overwrites,
-        reason='Very secret business.'
+        reason='Very secret business.',
     )
+
 
 @secret.command()
 @commands.guild_only()
@@ -81,7 +84,7 @@ async def emoji(ctx: commands.Context, emoji: discord.PartialEmoji, *roles: disc
     """
 
     # fetch the emoji asset and read it as bytes.
-    emoji_bytes = await emoji.url.read()
+    emoji_bytes = await emoji.read()
 
     # the key parameter here is `roles`, which controls
     # what roles are able to use the emoji.
@@ -89,7 +92,7 @@ async def emoji(ctx: commands.Context, emoji: discord.PartialEmoji, *roles: disc
         name=emoji.name,
         image=emoji_bytes,
         roles=roles,
-        reason='Very secret business.'
+        reason='Very secret business.',
     )
 
 
